Add Insertion Sort. This commit will add Insertion Sort, an in-place stable sorting algorithm.
Insertion Sort is useful for operating on small arrays due to it's
high speed, but it's asymptotic runtime complexity makes it an inferior
choice on large arrays.

Despite this, it's still a common sorting algorithm due to the ease of
implementation and the low memory requirements (a property of most
in-place sorting algorithms).